+++ packages/edit-post/src/components/header/header-toolbar/index.js
@@ -23,7 +23,6 @@ import { store as keyboardShortcutsStore } from '@wordpress/keyboard-shortcuts';
 /**
  * Internal dependencies
  */
-import TemplateTitle from '../template-title';
 import { store as editPostStore } from '../../../store';
 
 const preventDefault = ( event ) => {
@@ -67,7 +66,7 @@ function HeaderToolbar() {
 			showIconLabels: isFeatureActive( 'showIconLabels' ),
 			isListViewOpen: isListViewOpened(),
 			listViewShortcut: getShortcutRepresentation(
-				'core/edit-post/toggle-block-navigation'
+				'core/edit-post/toggle-list-view'
 			),
 		};
 	}, [] );
@@ -136,7 +135,8 @@ function HeaderToolbar() {
 					) }
 					showTooltip={ ! showIconLabels }
 				>
-					{ showIconLabels && __( 'Add' ) }
+					{ showIconLabels &&
+						( ! isInserterOpened ? __( 'Add' ) : __( 'Close' ) ) }
 				</ToolbarItem>
 				{ ( isWideViewport || ! showIconLabels ) && (
 					<>
@@ -164,8 +164,6 @@ function HeaderToolbar() {
 					</>
 				) }
 			</div>
-
-			<TemplateTitle />
 		</NavigableToolbar>
 	);
 }