+++ test/native/__mocks__/styleMock.js
@@ -107,9 +107,29 @@ module.exports = {
 		paddingRight: 10,
 	},
 	'block-types-list__column': {
-		padding: 10,
+		paddingLeft: 10,
+		paddingRight: 10,
 	},
 	floatingToolbar: {
 		height: 10,
+	},
+	searchFormPlaceholder: {
+		color: 'gray',
+	},
+	ripple: {
+		backgroundColor: 'white',
+	},
+	spacing: {
+		marginLeft: 6,
+		marginRight: 6,
+	},
+	arrow: {
+		color: 'red',
+	},
+	textInput: {
+		color: 'black',
+	},
+	buttonNoBg: {
+		color: 'orange',
 	},
 };